resolved merge conflicts

+++ UI/Views.Mac/TextBox.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 using AppInstall.Framework;
 using AppInstall.Graphics;
 
@@ -16,7 +16,7 @@ namespace AppInstall.UI
 
         public string Text { get { return nativeView.Text; } set { nativeView.Text = value; } }
         public string Placeholder { get { return nativeView.Placeholder; } set { nativeView.Placeholder = value; } }
-        public float FontSize { get { return nativeView.Font.PointSize; } set { nativeView.Font = nativeView.Font.WithSize(value); } }
+        public float FontSize { get { return (float)nativeView.Font.PointSize; } set { nativeView.Font = nativeView.Font.WithSize(value); } }
         public Color TextColor { get { return nativeView.TextColor.ToColor(); } set { nativeView.TextColor = value.ToUIColor(); } }
         public TextAlignment TextAlignment { get { return Abstraction.ToTextAlignment(nativeView.TextAlignment); } set { nativeView.TextAlignment = Abstraction.ToUITextAlignment(value); } }
         public bool Secure { get { return nativeView.SecureTextEntry; } set { nativeView.SecureTextEntry = value; } }
